Simplify element tracking for stateful components.
`Component` was keeping a redundnat currentElement field inside its
InternalData, named `element`.

This changes `Component:__update` to return a 'shouldContinueUpdate'
value to the reconciler, which tells the reconciler whether to abort the
update. The reconciler uses this value to determine whether to set
currentElement, which deduplicates some code and fixes some tracking
bugs.