Fix updateChildren re-rentrancy issue (#301)
In certain circumstances it is possible for updateChildren in createReconicler to be re-rendered for a virtualNode while it is already happening causing two Roblox instances to be created for the same element.

For an example of how this happens, see the added unit test.

What happens in the test is:

The components are rendered the first time and ChildComponent renders a Frame
In the spawn in ChildComponent:didMount, firstTime is set to false
ChildComponent's Frame is unmounted and a new TextLabel is added
The DescendantAdded event is processed
ChildComponent rerenders, the old virtualNode for the Frame is still in it's children.
ChildComponent's Frame is unmounted again and a second TextLabel is added
This example is kind of a contrived, but this did happen in the UIBlox GridView but with UI events such as AbsoluteContentSize Changed instead of DescentantAdded.

Solution
Suspend the events of parent nodes when updating a component outside of the standard lifecycle. Suspending the parent events means that no change we make can cause a parent node to reconcile while we are also reconciling.