Further update children re-rentrancy problems (#315)
This fixes the problem where nodes lower down the Roact tree 
can change outside of their standard lifecycle and then cause a 
callback or event to a node higher up the tree. That component
can re-render while the node further down the tree is updating 
its children, causing duplicates of elements to be added.